+++ include/snt/puq/to_string.h
@@ -7,9 +7,9 @@
 namespace snt::puq {
 
     /** Format a PUQ result. */
-    extern std::string to_string(const puq::Result& value, int precision = std::cout.precision());
+    extern std::string to_string(const puq::Result& value, int precision = static_cast<int>(std::cout.precision()));
     /** Format a base-unit expression. */
-    extern std::string to_string(const puq::BaseUnits& value, int precision = std::cout.precision());
+    extern std::string to_string(const puq::BaseUnits& value, int precision = static_cast<int>(std::cout.precision()));
     /** Format dimensions using unit-format options. */
     extern std::string to_string(const puq::Dimensions& value, const puq::UnitFormat& format = puq::UnitFormat());
     /** Format an exponent using unit-format options. */
